Modify method immigration_v2 for speed
Previously the loop goes through each species in each cell and draws
from a binomial dist to determine the # of immigrants for that sp in the
focal cell. Now the loop goes through each species and draws from a
binomial dist to determine the # of immigrants for that sp in ALL LOCAL
CELLS. Another loop then goes through the immigrating individuals and
determine their destination. This change cuts the time needed for
immigration (the time-limiting step) in half.